add prepared serializer
