+++ src/serializer/passThrough.ts
@@ -1,0 +1,6 @@
+import { Serializer } from '../types/Serializer'
+
+export const passThrough: Serializer<string> = {
+  parse: (x) => x,
+  stringify: (x) => x
+}
