+++ src/types/Serializer.ts
@@ -1,0 +1,4 @@
+export type Serializer<T> = {
+  parse: (value: string) => T
+  stringify: (value: T) => string
+}
